moved timetable classes to package
