+++ common/src/main/java/com/freshdigitable/yttt/feature/timetable/FetchStreamUseCase.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.yttt
+package com.freshdigitable.yttt.feature.timetable
 
 interface FetchStreamUseCase {
     suspend operator fun invoke()
