+++ common/src/main/java/com/freshdigitable/yttt/feature/timetable/TimetableContextMenuSelector.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.yttt
+package com.freshdigitable.yttt.feature.timetable
 
 import com.freshdigitable.yttt.data.model.LiveVideo
