+++ feature-timetable-twitch/src/main/java/com/freshdigitable/yttt/feature/timetable/twitch/FetchTwitchStreamUseCase.kt
@@ -1,8 +1,8 @@
 package com.freshdigitable.yttt.feature.timetable.twitch
 
-import com.freshdigitable.yttt.FetchStreamUseCase
 import com.freshdigitable.yttt.data.AccountRepository
 import com.freshdigitable.yttt.data.TwitchLiveRepository
+import com.freshdigitable.yttt.feature.timetable.FetchStreamUseCase
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
