+++ feature-timetable-twitch/src/main/java/com/freshdigitable/yttt/feature/timetable/twitch/TimetableContextMenuDelegateForTwitch.kt
@@ -1,9 +1,9 @@
 package com.freshdigitable.yttt.feature.timetable.twitch
 
 import com.freshdigitable.yttt.LaunchAppWithUrlUseCase
-import com.freshdigitable.yttt.TimetableContextMenuSelector
-import com.freshdigitable.yttt.TimetableMenuItem
 import com.freshdigitable.yttt.data.model.LiveVideo
+import com.freshdigitable.yttt.feature.timetable.TimetableContextMenuSelector
+import com.freshdigitable.yttt.feature.timetable.TimetableMenuItem
 import javax.inject.Inject
 
 internal class TimetableContextMenuDelegateForTwitch @Inject constructor(
