+++ feature-timetable-youtube/src/main/java/com/freshdigitable/yttt/feature/timetable/youtube/TimetablePageDelegateYouTube.kt
@@ -1,9 +1,9 @@
 package com.freshdigitable.yttt.feature.timetable.youtube
 
-import com.freshdigitable.yttt.FetchTimetableItemSourceUseCase
 import com.freshdigitable.yttt.data.YouTubeRepository
 import com.freshdigitable.yttt.data.model.LiveVideo
 import com.freshdigitable.yttt.data.model.toLiveVideo
+import com.freshdigitable.yttt.feature.timetable.FetchTimetableItemSourceUseCase
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
