+++ lib/src/main/java/com/freshdigitable/yttt/feature/timetable/TimetablePageDelegate.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.yttt
+package com.freshdigitable.yttt.feature.timetable
 
 import com.freshdigitable.yttt.compose.TabData
 import com.freshdigitable.yttt.data.SettingRepository
